+++ Ethar.GeoPose.UnitTests/BasicYawPitchRollSduTests.cs
@@ -10,9 +10,9 @@ namespace Ethar.GeoPose.UnitTests
 #if !UNITY_EDITOR && APITEST
         [TestCase(45, 45, 10)]
         [TestCase(0, 90, 100)]
-        [TestCase(45, 45, 10.12f)]
+        [TestCase(45, 45, 10.12)]
         [TestCase(-15, -45, -10)]
-        public async Task CanCorrectlyParseBasicYawPitchRollSdu(float lat, float lon, float height)
+        public async Task CanCorrectlyParseBasicYawPitchRollSdu(double lat, double lon, double height)
         {
             var sdu = await this.Client.ReadAsJsonAsync<BasicYawPitchRollSdu>($"{ApiEndpoints.BasicYprGet}?longitude={lon}&latitude={lat}&heightInMeters={height}");
             Assert.That(sdu.Position.Longitude, Is.EqualTo(lon));
@@ -21,8 +21,8 @@ namespace Ethar.GeoPose.UnitTests
         }
 #endif
 
-        [TestCase(45f, 45f, 5f, 12, 16, 8)]
-        public void CanCorrectlyDeserializeBasicYawPitchRollSdu(float lat, float lon, float h, float yaw, float pitch, float roll)
+        [TestCase(45, 45, 5, 12, 16, 8)]
+        public void CanCorrectlyDeserializeBasicYawPitchRollSdu(double lat, double lon, double h, double yaw, double pitch, double roll)
         {
             var json =
                 "{" +
@@ -50,8 +50,8 @@ namespace Ethar.GeoPose.UnitTests
             Assert.That(sdu.Angles.Roll, Is.EqualTo(roll));
         }
 
-        [TestCase(45.0f, 45.0f, 5, 12, 16, 8)]
-        public void CanCorrectlySerializeBasicYawPitchRollSdu(float lat, float lon, float h, float yaw, float pitch, float roll)
+        [TestCase(45.0, 45.0, 5, 12, 16, 8)]
+        public void CanCorrectlySerializeBasicYawPitchRollSdu(double lat, double lon, double h, double yaw, double pitch, double roll)
         {
             var sdu = new BasicYawPitchRollSdu(new DataTypes.YawPitchRollAngles() { Yaw = yaw, Pitch = pitch, Roll = roll }, new DataTypes.TangentPointPosition() { HeightInMeters = h, Latitude = lat, Longitude = lon });
 
@@ -77,7 +77,7 @@ namespace Ethar.GeoPose.UnitTests
         }
 
         [TestCase(45, 45, 5, 12, 16, 8)]
-        public void CanCorrectlyMakeARoundTripConversion(float lat, float lon, float h, float yaw, float pitch, float roll)
+        public void CanCorrectlyMakeARoundTripConversion(double lat, double lon, double h, double yaw, double pitch, double roll)
         {
             var sdu = new BasicYawPitchRollSdu(new DataTypes.YawPitchRollAngles() { Yaw = yaw, Pitch = pitch, Roll = roll }, new DataTypes.TangentPointPosition() { HeightInMeters = h, Latitude = lat, Longitude = lon });
 
@@ -88,8 +88,8 @@ namespace Ethar.GeoPose.UnitTests
             Assert.That(sdu, Is.EqualTo(converted));
         }
 
-        [TestCase(45f, 45f, 5f, 12, 16, 8)]
-        public void CanConvertBasicYawPitchRollSduToString(float lat, float lon, float h, float yaw, float pitch, float roll)
+        [TestCase(45, 45, 5, 12, 16, 8)]
+        public void CanConvertBasicYawPitchRollSduToString(double lat, double lon, double h, double yaw, double pitch, double roll)
         {
             var sdu = new BasicYawPitchRollSdu(new DataTypes.YawPitchRollAngles() { Yaw = yaw, Pitch = pitch, Roll = roll }, new DataTypes.TangentPointPosition() { HeightInMeters = h, Latitude = lat, Longitude = lon });
             var result = sdu.ToString();
